more updates

+++ cycle.go
@@ -18,7 +18,7 @@ func NewCycle(key string, values []string, opts ...ProxyOption) (*Cycle, error) 
 	return &Cycle{Proxy: *proxy, values: values}, nil
 }
 
-func (c *Cycle) Index() int {
+func (c *Cycle) Index() int64 {
 	return c.value()
 }
 
@@ -27,14 +27,14 @@ func (c *Cycle) Value() string {
 }
 
 func (c *Cycle) Next() (err error) {
-	value := (c.value() + 1) % len(c.values)
+	value := (c.value() + 1) % int64(len(c.values))
 	_, err = c.client.Set(c.ctx, c.key, value, c.expiresIn).Result()
 
 	return
 }
 
-func (c *Cycle) value() (v int) {
-	v, err := c.client.Get(c.ctx, c.key).Int()
+func (c *Cycle) value() (v int64) {
+	v, err := c.client.Get(c.ctx, c.key).Int64()
 	if err != nil && err != redis.Nil {
 		// TODO debug logging
 	}
